Modify the tests for ST_ShortestPath.
