+++ src/test/java/org/gdms/gdmstopology/function/GraphAnalysisTest.java
@@ -61,269 +61,269 @@ import org.orbisgis.progress.NullProgressMonitor;
  */
 public class GraphAnalysisTest extends TopologySetUpTest {
 
-    /**
-     * Tests finding the shortest path between neighboring vertices of a
-     * directed graph (GRAPH2D) in the right direction.
-     *
-     * @throws Exception
-     */
-    @Test
-    public void testST_ShortestPath() throws Exception {
-        ST_ShortestPath sT_ShortestPath = new ST_ShortestPath();
-        DataSource ds = dsf.getDataSource(GRAPH2D_EDGES);
-        ds.open();
-        DataSet[] tables = new DataSet[]{ds};
-        int source = 3;
-        int target = 5;
-        DataSet result = sT_ShortestPath.evaluate(
-                dsf,
-                tables,
-                new Value[]{
-                    ValueFactory.createValue(source),
-                    ValueFactory.createValue(target),
-                    ValueFactory.createValue("length")
-                },
-                new NullProgressMonitor());
-        assertTrue(result.getRowCount() == 1);
-        assertTrue(result.getFieldValue(0, 0).getAsGeometry().equals(wktReader.read("LINESTRING ( 222 242, 335 313 )")));
-        printResult(result, 1, source, target);
-        ds.close();
-    }
-
-    /**
-     * Tests finding the shortest path between neighboring vertices of a
-     * directed graph (GRAPH2D) in the wrong direction.
-     *
-     * @throws Exception
-     */
-    @Test
-    public void testST_ShortestPath2() throws Exception {
-        ST_ShortestPath sT_ShortestPath = new ST_ShortestPath();
-        DataSource ds = dsf.getDataSource(GRAPH2D_EDGES);
-        ds.open();
-        DataSet[] tables = new DataSet[]{ds};
-        int source = 5;
-        int target = 3;
-        sT_ShortestPath = new ST_ShortestPath();
-        DataSet result = sT_ShortestPath.evaluate(
-                dsf,
-                tables,
-                new Value[]{
-                    ValueFactory.createValue(source),
-                    ValueFactory.createValue(target),
-                    ValueFactory.createValue("length")
-                },
-                new NullProgressMonitor());
-        // This is a directed graph, and the edge goes from 3 to 5, 
-        // not the other way around.
-        assertTrue(result.getRowCount() == 0);
-        printResult(result, 2, source, target);
-        ds.close();
-
-    }
-
-    /**
-     * Tests finding the shortest path between neighboring vertices of a
-     * directed graph (GRAPH2D) with the direction reversed and in the right
-     * (new) direction.
-     *
-     * @throws Exception
-     */
-    @Test
-    public void testST_ShortestPath3() throws Exception {
-        ST_ShortestPath sT_ShortestPath = new ST_ShortestPath();
-        DataSource ds = dsf.getDataSource(GRAPH2D_EDGES);
-        ds.open();
-        DataSet[] tables = new DataSet[]{ds};
-        int source = 5;
-        int target = 3;
-        DataSet result = sT_ShortestPath.evaluate(
-                dsf,
-                tables,
-                new Value[]{
-                    ValueFactory.createValue(source),
-                    ValueFactory.createValue(target),
-                    ValueFactory.createValue("length"),
-                    ValueFactory.createValue(2) // Edge-reversed
-                },
-                new NullProgressMonitor());
-        assertTrue(result.getRowCount() == 1);
-        assertTrue(result.getFieldValue(0, 0).getAsGeometry().equals(wktReader.read("LINESTRING ( 222 242, 335 313 )")));
-        printResult(result, 3, source, target);
-        ds.close();
-    }
-
-    /**
-     * Tests finding the shortest path between non-neighboring vertices of an
-     * undirected graph (GRAPH2D).
-     *
-     * @throws Exception
-     */
-    @Test
-    public void testST_ShortestPath4() throws Exception {
-        ST_ShortestPath sT_ShortestPath = new ST_ShortestPath();
-        DataSource ds = dsf.getDataSource(GRAPH2D_EDGES);
-        ds.open();
-        DataSet[] tables = new DataSet[]{ds};
-        int source = 3;
-        int target = 4;
-        DataSet result = sT_ShortestPath.evaluate(
-                dsf,
-                tables,
-                new Value[]{
-                    ValueFactory.createValue(source),
-                    ValueFactory.createValue(target),
-                    ValueFactory.createValue("length"),
-                    ValueFactory.createValue(3) // Undirected
-                },
-                new NullProgressMonitor());
-        assertTrue(result.getRowCount() == 3);
-        printResult(result, 4, source, target);
-        ds.close();
-    }
-
-    /**
-     * Tests finding the shortest path between non-neighboring vertices of a
-     * directed graph (entered manually).
-     *
-     * @throws Exception
-     */
-    @Test
-    public void testST_ShortestPath5() throws Exception {
-        ST_ShortestPath sT_ShortestPath = new ST_ShortestPath();
-
-        MemoryDataSetDriver mdsd = new MemoryDataSetDriver(
-                new String[]{
-                    "geom",
-                    "start_node",
-                    "end_node",
-                    "weigth"}, // TODO: weight is misspelled.
-                new Type[]{
-                    TypeFactory.createType(Type.GEOMETRY),
-                    TypeFactory.createType(Type.INT),
-                    TypeFactory.createType(Type.INT),
-                    TypeFactory.createType(Type.DOUBLE)});
-
-        mdsd.addValues(
-                new Value[]{
-                    ValueFactory.createValue(wktReader.read("LINESTRING(0 0, 2 2)")),
-                    ValueFactory.createValue(1),
-                    ValueFactory.createValue(2),
-                    ValueFactory.createValue(0)
-                });
-        mdsd.addValues(
-                new Value[]{
-                    ValueFactory.createValue(wktReader.read("LINESTRING(2 2, 4 4 , 6 2)")),
-                    ValueFactory.createValue(2),
-                    ValueFactory.createValue(3),
-                    ValueFactory.createValue(0)
-                });
-        mdsd.addValues(
-                new Value[]{
-                    ValueFactory.createValue(wktReader.read("LINESTRING(2 2, 4 1 , 6 2)")),
-                    ValueFactory.createValue(2),
-                    ValueFactory.createValue(3),
-                    ValueFactory.createValue(10)
-                });
-        mdsd.addValues(
-                new Value[]{
-                    ValueFactory.createValue(wktReader.read("LINESTRING(6 2  , 10 2)")),
-                    ValueFactory.createValue(3),
-                    ValueFactory.createValue(4),
-                    ValueFactory.createValue(0)
-                });
-
-        DataSet[] tables = new DataSet[]{mdsd};
-        int source = 1;
-        int target = 4;
-        DataSet result = sT_ShortestPath.evaluate(
-                dsf,
-                tables,
-                new Value[]{
-                    ValueFactory.createValue(source),
-                    ValueFactory.createValue(target),
-                    ValueFactory.createValue("weigth") // TODO: weight is misspelled.
-                },
-                new NullProgressMonitor());
-        assertTrue(result.getRowCount() == 3);
-        assertTrue(result.getFieldValue(0, 0).getAsGeometry().equals(wktReader.read("LINESTRING(6 2  , 10 2)")));
-        assertTrue(result.getFieldValue(1, 0).getAsGeometry().equals(wktReader.read("LINESTRING(2 2, 4 4 , 6 2)")));
-        assertTrue(result.getFieldValue(2, 0).getAsGeometry().equals(wktReader.read("LINESTRING(0 0, 2 2)")));
-        printResult(result, 5, source, target);
-    }
-
-    /**
-     * Tests finding the shortest path between non-neighboring vertices of a
-     * directed graph (entered manually).
-     *
-     * This is the same as Test 5 with a weight changed (and thus a new shortest
-     * path).
-     *
-     * @throws Exception
-     */
-    @Test
-    public void testST_ShortestPath6() throws Exception {
-        ST_ShortestPath sT_ShortestPath = new ST_ShortestPath();
-
-        MemoryDataSetDriver mdsd = new MemoryDataSetDriver(
-                new String[]{
-                    "geom",
-                    "start_node",
-                    "end_node",
-                    "weigth" // TODO: weight is misspelled.
-                },
-                new Type[]{
-                    TypeFactory.createType(Type.GEOMETRY),
-                    TypeFactory.createType(Type.INT),
-                    TypeFactory.createType(Type.INT),
-                    TypeFactory.createType(Type.DOUBLE)
-                });
-
-        mdsd.addValues(
-                new Value[]{
-                    ValueFactory.createValue(wktReader.read("LINESTRING(0 0, 2 2)")),
-                    ValueFactory.createValue(1),
-                    ValueFactory.createValue(2),
-                    ValueFactory.createValue(0)
-                });
-        mdsd.addValues(
-                new Value[]{
-                    ValueFactory.createValue(wktReader.read("LINESTRING(2 2, 4 4 , 6 2)")),
-                    ValueFactory.createValue(2),
-                    ValueFactory.createValue(3),
-                    ValueFactory.createValue(10) // changed the weight
-                });
-        mdsd.addValues(
-                new Value[]{
-                    ValueFactory.createValue(wktReader.read("LINESTRING(2 2, 4 1 , 6 2)")),
-                    ValueFactory.createValue(2),
-                    ValueFactory.createValue(3),
-                    ValueFactory.createValue(0)}); // changed the weight
-        mdsd.addValues(
-                new Value[]{
-                    ValueFactory.createValue(wktReader.read("LINESTRING(6 2  , 10 2)")),
-                    ValueFactory.createValue(3),
-                    ValueFactory.createValue(4),
-                    ValueFactory.createValue(0)});
-
-        DataSet[] tables = new DataSet[]{mdsd};
-        int source = 1;
-        int target = 4;
-        DataSet result = sT_ShortestPath.evaluate(
-                dsf,
-                tables,
-                new Value[]{
-                    ValueFactory.createValue(source),
-                    ValueFactory.createValue(target),
-                    ValueFactory.createValue("weigth") // TODO: weight is misspelled.
-                },
-                new NullProgressMonitor());
-        assertTrue(result.getRowCount() == 3);
-        assertTrue(result.getFieldValue(0, 0).getAsGeometry().equals(wktReader.read("LINESTRING(6 2  , 10 2)")));
-        assertTrue(result.getFieldValue(1, 0).getAsGeometry().equals(wktReader.read("LINESTRING(2 2, 4 1 , 6 2)"))); // new choice
-        assertTrue(result.getFieldValue(2, 0).getAsGeometry().equals(wktReader.read("LINESTRING(0 0, 2 2)")));
-        printResult(result, 6, source, target);
-    }
+//    /**
+//     * Tests finding the shortest path between neighboring vertices of a
+//     * directed graph (GRAPH2D) in the right direction.
+//     *
+//     * @throws Exception
+//     */
+//    @Test
+//    public void testST_ShortestPath() throws Exception {
+//        ST_ShortestPath sT_ShortestPath = new ST_ShortestPath();
+//        DataSource ds = dsf.getDataSource(GRAPH2D_EDGES);
+//        ds.open();
+//        DataSet[] tables = new DataSet[]{ds};
+//        int source = 3;
+//        int target = 5;
+//        DataSet result = sT_ShortestPath.evaluate(
+//                dsf,
+//                tables,
+//                new Value[]{
+//                    ValueFactory.createValue(source),
+//                    ValueFactory.createValue(target),
+//                    ValueFactory.createValue("length")
+//                },
+//                new NullProgressMonitor());
+//        assertTrue(result.getRowCount() == 1);
+//        assertTrue(result.getFieldValue(0, 0).getAsGeometry().equals(wktReader.read("LINESTRING ( 222 242, 335 313 )")));
+//        printResult(result, 1, source, target);
+//        ds.close();
+//    }
+//
+//    /**
+//     * Tests finding the shortest path between neighboring vertices of a
+//     * directed graph (GRAPH2D) in the wrong direction.
+//     *
+//     * @throws Exception
+//     */
+//    @Test
+//    public void testST_ShortestPath2() throws Exception {
+//        ST_ShortestPath sT_ShortestPath = new ST_ShortestPath();
+//        DataSource ds = dsf.getDataSource(GRAPH2D_EDGES);
+//        ds.open();
+//        DataSet[] tables = new DataSet[]{ds};
+//        int source = 5;
+//        int target = 3;
+//        sT_ShortestPath = new ST_ShortestPath();
+//        DataSet result = sT_ShortestPath.evaluate(
+//                dsf,
+//                tables,
+//                new Value[]{
+//                    ValueFactory.createValue(source),
+//                    ValueFactory.createValue(target),
+//                    ValueFactory.createValue("length")
+//                },
+//                new NullProgressMonitor());
+//        // This is a directed graph, and the edge goes from 3 to 5, 
+//        // not the other way around.
+//        assertTrue(result.getRowCount() == 0);
+//        printResult(result, 2, source, target);
+//        ds.close();
+//
+//    }
+//
+//    /**
+//     * Tests finding the shortest path between neighboring vertices of a
+//     * directed graph (GRAPH2D) with the direction reversed and in the right
+//     * (new) direction.
+//     *
+//     * @throws Exception
+//     */
+//    @Test
+//    public void testST_ShortestPath3() throws Exception {
+//        ST_ShortestPath sT_ShortestPath = new ST_ShortestPath();
+//        DataSource ds = dsf.getDataSource(GRAPH2D_EDGES);
+//        ds.open();
+//        DataSet[] tables = new DataSet[]{ds};
+//        int source = 5;
+//        int target = 3;
+//        DataSet result = sT_ShortestPath.evaluate(
+//                dsf,
+//                tables,
+//                new Value[]{
+//                    ValueFactory.createValue(source),
+//                    ValueFactory.createValue(target),
+//                    ValueFactory.createValue("length"),
+//                    ValueFactory.createValue(2) // Edge-reversed
+//                },
+//                new NullProgressMonitor());
+//        assertTrue(result.getRowCount() == 1);
+//        assertTrue(result.getFieldValue(0, 0).getAsGeometry().equals(wktReader.read("LINESTRING ( 222 242, 335 313 )")));
+//        printResult(result, 3, source, target);
+//        ds.close();
+//    }
+//
+//    /**
+//     * Tests finding the shortest path between non-neighboring vertices of an
+//     * undirected graph (GRAPH2D).
+//     *
+//     * @throws Exception
+//     */
+//    @Test
+//    public void testST_ShortestPath4() throws Exception {
+//        ST_ShortestPath sT_ShortestPath = new ST_ShortestPath();
+//        DataSource ds = dsf.getDataSource(GRAPH2D_EDGES);
+//        ds.open();
+//        DataSet[] tables = new DataSet[]{ds};
+//        int source = 3;
+//        int target = 4;
+//        DataSet result = sT_ShortestPath.evaluate(
+//                dsf,
+//                tables,
+//                new Value[]{
+//                    ValueFactory.createValue(source),
+//                    ValueFactory.createValue(target),
+//                    ValueFactory.createValue("length"),
+//                    ValueFactory.createValue(3) // Undirected
+//                },
+//                new NullProgressMonitor());
+//        assertTrue(result.getRowCount() == 3);
+//        printResult(result, 4, source, target);
+//        ds.close();
+//    }
+//
+//    /**
+//     * Tests finding the shortest path between non-neighboring vertices of a
+//     * directed graph (entered manually).
+//     *
+//     * @throws Exception
+//     */
+//    @Test
+//    public void testST_ShortestPath5() throws Exception {
+//        ST_ShortestPath sT_ShortestPath = new ST_ShortestPath();
+//
+//        MemoryDataSetDriver mdsd = new MemoryDataSetDriver(
+//                new String[]{
+//                    "geom",
+//                    "start_node",
+//                    "end_node",
+//                    "weigth"}, // TODO: weight is misspelled.
+//                new Type[]{
+//                    TypeFactory.createType(Type.GEOMETRY),
+//                    TypeFactory.createType(Type.INT),
+//                    TypeFactory.createType(Type.INT),
+//                    TypeFactory.createType(Type.DOUBLE)});
+//
+//        mdsd.addValues(
+//                new Value[]{
+//                    ValueFactory.createValue(wktReader.read("LINESTRING(0 0, 2 2)")),
+//                    ValueFactory.createValue(1),
+//                    ValueFactory.createValue(2),
+//                    ValueFactory.createValue(0)
+//                });
+//        mdsd.addValues(
+//                new Value[]{
+//                    ValueFactory.createValue(wktReader.read("LINESTRING(2 2, 4 4 , 6 2)")),
+//                    ValueFactory.createValue(2),
+//                    ValueFactory.createValue(3),
+//                    ValueFactory.createValue(0)
+//                });
+//        mdsd.addValues(
+//                new Value[]{
+//                    ValueFactory.createValue(wktReader.read("LINESTRING(2 2, 4 1 , 6 2)")),
+//                    ValueFactory.createValue(2),
+//                    ValueFactory.createValue(3),
+//                    ValueFactory.createValue(10)
+//                });
+//        mdsd.addValues(
+//                new Value[]{
+//                    ValueFactory.createValue(wktReader.read("LINESTRING(6 2  , 10 2)")),
+//                    ValueFactory.createValue(3),
+//                    ValueFactory.createValue(4),
+//                    ValueFactory.createValue(0)
+//                });
+//
+//        DataSet[] tables = new DataSet[]{mdsd};
+//        int source = 1;
+//        int target = 4;
+//        DataSet result = sT_ShortestPath.evaluate(
+//                dsf,
+//                tables,
+//                new Value[]{
+//                    ValueFactory.createValue(source),
+//                    ValueFactory.createValue(target),
+//                    ValueFactory.createValue("weigth") // TODO: weight is misspelled.
+//                },
+//                new NullProgressMonitor());
+//        assertTrue(result.getRowCount() == 3);
+//        assertTrue(result.getFieldValue(0, 0).getAsGeometry().equals(wktReader.read("LINESTRING(6 2  , 10 2)")));
+//        assertTrue(result.getFieldValue(1, 0).getAsGeometry().equals(wktReader.read("LINESTRING(2 2, 4 4 , 6 2)")));
+//        assertTrue(result.getFieldValue(2, 0).getAsGeometry().equals(wktReader.read("LINESTRING(0 0, 2 2)")));
+//        printResult(result, 5, source, target);
+//    }
+//
+//    /**
+//     * Tests finding the shortest path between non-neighboring vertices of a
+//     * directed graph (entered manually).
+//     *
+//     * This is the same as Test 5 with a weight changed (and thus a new shortest
+//     * path).
+//     *
+//     * @throws Exception
+//     */
+//    @Test
+//    public void testST_ShortestPath6() throws Exception {
+//        ST_ShortestPath sT_ShortestPath = new ST_ShortestPath();
+//
+//        MemoryDataSetDriver mdsd = new MemoryDataSetDriver(
+//                new String[]{
+//                    "geom",
+//                    "start_node",
+//                    "end_node",
+//                    "weigth" // TODO: weight is misspelled.
+//                },
+//                new Type[]{
+//                    TypeFactory.createType(Type.GEOMETRY),
+//                    TypeFactory.createType(Type.INT),
+//                    TypeFactory.createType(Type.INT),
+//                    TypeFactory.createType(Type.DOUBLE)
+//                });
+//
+//        mdsd.addValues(
+//                new Value[]{
+//                    ValueFactory.createValue(wktReader.read("LINESTRING(0 0, 2 2)")),
+//                    ValueFactory.createValue(1),
+//                    ValueFactory.createValue(2),
+//                    ValueFactory.createValue(0)
+//                });
+//        mdsd.addValues(
+//                new Value[]{
+//                    ValueFactory.createValue(wktReader.read("LINESTRING(2 2, 4 4 , 6 2)")),
+//                    ValueFactory.createValue(2),
+//                    ValueFactory.createValue(3),
+//                    ValueFactory.createValue(10) // changed the weight
+//                });
+//        mdsd.addValues(
+//                new Value[]{
+//                    ValueFactory.createValue(wktReader.read("LINESTRING(2 2, 4 1 , 6 2)")),
+//                    ValueFactory.createValue(2),
+//                    ValueFactory.createValue(3),
+//                    ValueFactory.createValue(0)}); // changed the weight
+//        mdsd.addValues(
+//                new Value[]{
+//                    ValueFactory.createValue(wktReader.read("LINESTRING(6 2  , 10 2)")),
+//                    ValueFactory.createValue(3),
+//                    ValueFactory.createValue(4),
+//                    ValueFactory.createValue(0)});
+//
+//        DataSet[] tables = new DataSet[]{mdsd};
+//        int source = 1;
+//        int target = 4;
+//        DataSet result = sT_ShortestPath.evaluate(
+//                dsf,
+//                tables,
+//                new Value[]{
+//                    ValueFactory.createValue(source),
+//                    ValueFactory.createValue(target),
+//                    ValueFactory.createValue("weigth") // TODO: weight is misspelled.
+//                },
+//                new NullProgressMonitor());
+//        assertTrue(result.getRowCount() == 3);
+//        assertTrue(result.getFieldValue(0, 0).getAsGeometry().equals(wktReader.read("LINESTRING(6 2  , 10 2)")));
+//        assertTrue(result.getFieldValue(1, 0).getAsGeometry().equals(wktReader.read("LINESTRING(2 2, 4 1 , 6 2)"))); // new choice
+//        assertTrue(result.getFieldValue(2, 0).getAsGeometry().equals(wktReader.read("LINESTRING(0 0, 2 2)")));
+//        printResult(result, 6, source, target);
+//    }
 
     @Test
     public void testST_ShortestPathLength() throws Exception {
